+++ plugins/csv-parser/src/node.d.ts
@@ -1,13 +1,13 @@
-import { DataParser } from './dataParser';
-import { Options } from 'csv-parse/.';
+import { DataParser } from './dataParser'
+import { Options } from 'csv-parse/.'
 interface dataType {
-    series: {
-        name: string;
-        values: string[] | number[];
-    }[];
+	series: {
+		name: string
+		values: string[] | number[]
+	}[]
 }
 export declare class CSVParser extends DataParser {
-    parse(input: string, options?: Options): Promise<dataType | null>;
-    readCSVFile(fileName: string): string;
+	parse(input: string, options?: Options): Promise<dataType | null>
+	readCSVFile(fileName: string): string
 }
-export default CSVParser;
+export default CSVParser